+++ lib/Model/Reject.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * RedirectAllOf
+ * Reject
  *
  * PHP version 5
  *
@@ -28,19 +28,18 @@
  */
 
 namespace OpenAPI\Client\Model;
-
-use \ArrayAccess;
 use \OpenAPI\Client\ObjectSerializer;
 
 /**
- * RedirectAllOf Class Doc Comment
+ * Reject Class Doc Comment
  *
  * @category Class
+ * @description The &#x60;Reject&#x60; command blocks an incoming Call. Using &#x60;Reject&#x60; is the only way to prevent FreeClimb from answering a Call. Any other response will result in an answered Call and your account will be billed.
  * @package  OpenAPI\Client
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
  */
-class RedirectAllOf implements ModelInterface, ArrayAccess
+class Reject extends PerclCommand 
 {
     const DISCRIMINATOR = null;
 
@@ -49,7 +48,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
       *
       * @var string
       */
-    protected static $openAPIModelName = 'Redirect_allOf';
+    protected static $openAPIModelName = 'Reject';
 
     /**
       * Array of property to type mappings. Used for (de)serialization
@@ -57,7 +56,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $openAPITypes = [
-        'action_url' => 'string'
+        'reason' => 'string'
     ];
 
     /**
@@ -66,7 +65,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $openAPIFormats = [
-        'action_url' => null
+        'reason' => null
     ];
 
     /**
@@ -76,7 +75,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public static function openAPITypes()
     {
-        return self::$openAPITypes;
+        return self::$openAPITypes + parent::openAPITypes();
     }
 
     /**
@@ -86,7 +85,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public static function openAPIFormats()
     {
-        return self::$openAPIFormats;
+        return self::$openAPIFormats + parent::openAPIFormats();
     }
 
     /**
@@ -96,7 +95,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $attributeMap = [
-        'action_url' => 'actionUrl'
+        'reason' => 'reason'
     ];
 
     /**
@@ -105,7 +104,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $setters = [
-        'action_url' => 'setActionUrl'
+        'reason' => 'setReason'
     ];
 
     /**
@@ -114,7 +113,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $getters = [
-        'action_url' => 'getActionUrl'
+        'reason' => 'getReason'
     ];
 
     /**
@@ -125,7 +124,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public static function attributeMap()
     {
-        return self::$attributeMap;
+        return parent::attributeMap() + self::$attributeMap;
     }
 
     /**
@@ -135,7 +134,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public static function setters()
     {
-        return self::$setters;
+        return parent::setters() + self::$setters;
     }
 
     /**
@@ -145,7 +144,7 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public static function getters()
     {
-        return self::$getters;
+        return parent::getters() + self::$getters;
     }
 
     /**
@@ -162,12 +161,6 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
 
     
 
-    /**
-     * Associative array for storing property values
-     *
-     * @var mixed[]
-     */
-    protected $container = [];
 
     /**
      * Constructor
@@ -177,7 +170,9 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['action_url'] = isset($data['action_url']) ? $data['action_url'] : null;
+        parent::__construct($data);
+
+        $this->container['reason'] = isset($data['reason']) ? $data['reason'] : null;
     }
 
     /**
@@ -187,11 +182,8 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
      */
     public function listInvalidProperties()
     {
-        $invalidProperties = [];
+        $invalidProperties = parent::listInvalidProperties();
 
-        if ($this->container['action_url'] === null) {
-            $invalidProperties[] = "'action_url' can't be null";
-        }
         return $invalidProperties;
     }
 
@@ -208,25 +200,25 @@ class RedirectAllOf implements ModelInterface, ArrayAccess
 
 
     /**
-     * Gets action_url
+     * Gets reason
      *
-     * @return string
+     * @return string|null
      */
-    public function getActionUrl()
+    public function getReason()
     {
-        return $this->container['action_url'];
+        return $this->container['reason'];
     }
 
     /**
-     * Sets action_url
+     * Sets reason
      *
-     * @param string $action_url URL to request a new PerCL script to continue with the current Call's processing. When `Redirect` invokes the `actionUrl`, an `inbound` Webhook is sent. This request therefore looks identical to the initial request (made to the `voiceUrl` of the number that was called) for an inbound Call.
+     * @param string|null $reason Reason for the rejection. This can be any string value. In general, applications should use a set of enumerated values that are predefined to cover all exit points of the call flows for the given application.
      *
      * @return $this
      */
-    public function setActionUrl($action_url)
+    public function setReason($reason)
     {
-        $this->container['action_url'] = $action_url;
+        $this->container['reason'] = $reason;
 
         return $this;
     }